+++ tests/conftest.py
@@ -2,11 +2,13 @@
 
 import inspect
 import os
+from collections import namedtuple
 from operator import itemgetter
 
 import pytest
 
 from wemake_python_styleguide import violations
+from wemake_python_styleguide.options.config import Configuration
 from wemake_python_styleguide.violations.base import (
     ASTViolation,
     BaseViolation,
@@ -70,3 +72,27 @@ def absolute_path():
         return os.path.join(dirname, *files)
 
     return factory
+
+
+@pytest.fixture(scope='session')
+def options():
+    """Returns the options builder."""
+    default_values = {
+        option.long_option_name[2:].replace('-', '_'): option.default
+        for option in Configuration.options
+    }
+
+    Options = namedtuple('options', default_values.keys())
+
+    def factory(**kwargs):
+        final_options = default_values.copy()
+        final_options.update(kwargs)
+        return Options(**final_options)
+
+    return factory
+
+
+@pytest.fixture(scope='session')
+def default_options(options):
+    """Returns the default options."""
+    return options()
