+++ tests/test_checkers/test_module_names.py
@@ -16,9 +16,10 @@ from wemake_python_styleguide.violations import naming
     ('123py.py', naming.WrongModuleNamePatternViolation),
     ('version_1.py', naming.UnderScoredNumberNameViolation),
 ])
-def test_module_names(filename, error):
+def test_module_names(filename, error, default_options):
     """Ensures that checker works with module names."""
     module = ast.parse('')
+    Checker.parse_options(default_options)
     checker = Checker(tree=module, file_tokens=[], filename=filename)
     _, _, error_text, _ = next(checker.run())
     error_code = int(error_text[1:4])
