+++ tests/test_visitors/conftest.py
@@ -1,16 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from collections import namedtuple
 from typing import Sequence
 
 import pytest
 
-from wemake_python_styleguide.options.config import Configuration
 from wemake_python_styleguide.visitors.base import BaseVisitor
-
-
-def _to_dest_option(long_option_name: str) -> str:
-    return long_option_name[2:].replace('-', '_')
 
 
 @pytest.fixture(scope='session')
@@ -23,27 +17,3 @@ def assert_errors():
             assert error.code == errors[index].code
 
     return factory
-
-
-@pytest.fixture(scope='session')
-def options():
-    """Returns the options builder."""
-    default_values = {
-        _to_dest_option(option.long_option_name): option.default
-        for option in Configuration.options
-    }
-
-    Options = namedtuple('options', default_values.keys())
-
-    def factory(**kwargs):
-        final_options = default_values.copy()
-        final_options.update(kwargs)
-        return Options(**final_options)
-
-    return factory
-
-
-@pytest.fixture(scope='session')
-def default_options(options):
-    """Returns the default options."""
-    return options()
